不再支持 Android embedding v1
