+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.0.0-iOS-NoPay
 
 * nullsafety
+* 不再支持 Android embedding v1
 
 ## 2.0.0-iOS-NoPay+1
